requirements.txt
dash==2.9.3
dash-bootstrap-components==1.3.1
flask-caching==1.10.1
SQLAlchemy==1.4.46
pandas==1.5.3
numpy==1.24.3
yfinance==0.2.24
plotly==5.15.0
python-dotenv==1.0.0
aiohttp==3.8.4